fix: enable special SQL
